requirements.txt
streamlit
langchain-community
langchain-experimental
langchain-huggingface
faiss-cpu
pdfplumber
ollama